Вынес отдельно в models.dart классы Product и Category
