+++ lib/src/features/menu/view/widgets/models.dart
@@ -1,0 +1,35 @@
+class Product {
+  final int id;
+  final String name;
+  final String description;
+  final String imageUrl;
+  final String price;
+
+  Product({
+    required this.id,
+    required this.name,
+    required this.description,
+    required this.imageUrl,
+    required this.price,
+  });
+
+  factory Product.fromJson(Map<String, dynamic> json) {
+    return Product(
+      id: json['id'] as int,
+      name: json['name'] as String,
+      description: json['description'] as String,
+      imageUrl: json['imageUrl'] as String,
+      price: json['prices'][0]['value'] as String,
+    );
+  }
+}
+
+class Category {
+  final String name;
+  final List<Product> products;
+
+  Category({
+    required this.name,
+    required this.products,
+  });
+}
